separated cj_random and ObjectBlueprints
